requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
Deprecated==1.2.13
Django==3.2
django-oauth-toolkit==2.0.0
djangorestframework==3.12.0
drf-yasg==1.20.0
elasticsearch==7.6.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
jwcrypto==1.3.1
MarkupSafe==2.1.1
nltk==3.7
oauthlib==3.2.0
packaging==21.3
polyglot==16.7.4
psycopg2-binary==2.9.3
pycld2==0.41
pycparser==2.21
PyICU==2.9
PyJWT==2.3.0
pyparsing==3.0.9
PyPDF2==2.1.0
pytz==2022.1
regex==2022.6.2
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1